Change is_public logic
Previously, Universe privacy was checked first. If it was private, then
any content within it was private. If it was public, only then would
the content's privacy be checked.

Now, both values are checked in an OR comparison. If either the universe
or content privacy is "public," then the content is public.