Shrink images to the size they are actually used
Also adjust pixel dimensions very slightly so that all
images are the exact same dimensions. We're talkin'
one- or two-pixel changes, but the misalignment is
noticeable if you look closely.

Finally, the pixel density of the images are adjusted
as well. The change is not noticeable, but results in
much smaller file sizes.

These changes result in a visual improvement. Not only
does it fix the aforementioned aspect ratio mismatches,
it improves the quality of the scaled-down images.
I guess the browser did a poor job scaling the images
down, because more detail is visible in the scaled-down
versions.