Fix checkboxes to use materialize checkboxes
Before there was a hack in place to not use materialize checkboxes
because the actual checkbox didn't show up.

This re-arranges all the checkboxes I could find to work with
materialize, thus allowing the hack to be safely removed.

The forum (namely preferences) continue to use the hacky checkbox, but
that hack is maintained separately and still seems to work with this
change.