Image Improvements
Optimised all 38 Images in card headers, which was mostly jpg using TinyJPG and TinyPNG. 
Was a mixture of JPG and JPEG, JPG that actually were PNG's, so these fully switched over to JPG, so it's size was more realistic.
Total file size of Original files was 8.45mb, the optimised files is 3.60mb. No significant image issues were found that weren't already there.