Image improve for Logos
The files in Logos was optimised using TinyPNG. Then Converted to WebP with GIMP to preserve transparency. 
Total file size of Original files was 485kb, the optimised files is 164kb, the WebP files is 122kb. 
No significant image issues were found that weren't already there. 
While still an improvement overall, WebP, is hit or miss on PNG's that use transparency, when already optimised.